requirements.txt
fastapi==0.112.2
uvicorn==0.30.6
mongoengine==0.29.0
passlib==1.7.4
bcrypt==4.0.1
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
pydantic[email]==2.9.1